Just to get some function to the socket-package
The socket-package is not functional due to a strange mixture of different development strategies. A lot of client-site socket-logic moved to angular-directives. These right know have problems to communicate with the module-scope. (The first approach with most of the logic in the controller is much easier to understand, the second approach is overkill)

This index.html brings at least some function to the module. Most of the rest has to be done in the angular directive-file.